+++ mpqp/execution/providers/atos.py
@@ -513,7 +513,7 @@ def extract_state_vector_result(
         nb_qubits = job.circuit.nb_qubits
 
     nb_states = 2**nb_qubits
-    amplitudes = np.zeros(nb_states, np.complex64)
+    amplitudes = np.zeros(nb_states, np.complex128)
     probas = np.zeros(nb_states, np.float32)
     for sample in myqlm_result:
         state = sample.state.int
@@ -612,7 +612,7 @@ def extract_observable_result(
             ExpectationMeasure(
                 targets=list(range(nb_qubits)),
                 observable=Observable(
-                    np.zeros((2**nb_qubits, 2**nb_qubits), dtype=np.complex64)
+                    np.zeros((2**nb_qubits, 2**nb_qubits), dtype=np.complex128)
                 ),
                 shots=nb_shots,
             ),
